requirements/base.txt
asgiref==3.2.10
astroid==2.2.5
attrs==19.3.0
bleach==3.3.0
bleach-whitelist==0.0.10
certifi==2019.11.28
coverage==5.0
Django==3.1.8
django-ckeditor==5.9.0
django-comments-dab==2.6.1
django-crispy-forms==1.9.2
django-etc==1.2.0
django-hitcount==1.3.3
django-js-asset==1.2.2
django-markup==1.4.1
django-markupfield==2.0.0
django-meta==1.7.0
django-taggit==1.2.0
django-taggit-autosuggest==0.3.8
django-webp==2.0.0
dnspython==1.16.0
docutils==0.15.2
fancycompleter==0.8
html5lib==1.0.1
httpagentparser==1.9.0
idna==2.8
importlib-metadata==1.3.0
isort==4.3.21
lazy-object-proxy==1.4.2
Markdown==3.1.1
mccabe==0.6.1
more-itertools==8.0.2
mysqlclient==1.4.6
packaging==19.2
Pillow==8.1.1
# pkg-resources==0.0.0
pluggy==0.13.1
py==1.10.0
py3-validate-email==0.2.0
pycodestyle==2.5.0
python-creole==1.3.2
pytz==2019.2
six==1.15.0
smartypants==2.0.1
sqlparse==0.3.0
textile==3.0.4
typed-ast==1.4.0
urllib3==1.25.8
wcwidth==0.1.7
webencodings==0.5.1
wmctrl==0.3
wrapt==1.11.2
zipp==0.6.0
